# docs/swagger.yaml
basePath: /api
definitions:
  web.BinaryGroupResponse:
    properties:
      name:
        type: string
      versions:
        items:
          $ref: '#/definitions/web.BinaryResponse'
        type: array
    type: object
  web.BinaryResponse:
    description: Binary file information
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      path:
        type: string
      playbook:
        type: string
      updated_at:
        type: string
      version:
        type: string
    type: object
  web.ErrorResponse:
    description: Error response from the API
    properties:
      error:
        type: string
    type: object
info:
  contact: {}
  description: API for generating and managing binary files from playbooks
  title: Spage API
  version: "1.0"
paths:
  /binaries:
    get:
      description: Get a list of all generated binaries
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/web.BinaryResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: List all binaries
      tags:
      - binaries
  /binaries/{id}/download:
    get:
      description: Download a specific version of a binary file
      parameters:
      - description: Binary ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: Download specific binary version
      tags:
      - binaries
  /binaries/grouped:
    get:
      description: Get a list of binaries grouped by name
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/web.BinaryGroupResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: List grouped binaries
      tags:
      - binaries
  /generate:
    post:
      description: Generate a binary from a playbook
      parameters:
      - description: Binary name
        in: body
        name: name
        required: true
        schema:
          type: string
      - description: Playbook content
        in: body
        name: playbook
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: Generate binary from playbook
      tags:
      - generate
swagger: "2.0"